requirements.txt
anyio==3.6.2
appdirs==1.4.4
asgiref==3.3.4
attrs==21.2.0
boto3==1.17.67
botocore==1.20.67
cachetools==4.2.4
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
construct==2.10.68
construct-typing==0.5.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.2
django-allauth==0.44.0
django-amplitude==0.7.0
django-classy-tags==2.0.0
django-cors-headers==3.11.0
django-rest-swagger==2.2.0
django-storages==1.11.1
django-taggit==1.4.0
django-taggit-templatetags==0.2.5
django-taggit-templatetags2==1.6.1
django-templatetag-sugar==1.0
djangorestframework==3.13.1
drf-yasg==1.21.5
filelock==3.0.12
gunicorn==20.1.0
h11==0.14.0
h2==3.2.0
hackle-sdk==2.1.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.16.3
httpx==0.23.3
hyperframe==5.2.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
jsonalias==0.1.1
jsonschema==3.2.0
MarkupSafe==2.1.2
mmh3==2.5.1
mysql-client==0.0.1
mysqlclient==2.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==23.0
Pillow==8.2.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
pyOpenSSL==21.0.0
pyrsistent==0.18.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rethinkdb==2.4.8
rfc3986==1.5.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.4.2
simplejson==3.18.3
six==1.15.0
sniffio==1.2.0
solana==0.29.1
solders==0.14.4
sqlparse==0.4.1
thinker==1.1.1
types-cachetools==4.2.10
typing-extensions==4.5.0
ua-parser==0.10.0
uritemplate==4.1.1
urllib3==1.26.4
user-agents==2.2.0
virtualenv==20.4.4
websockets==10.4
Wrapper==1.1.0b1
